+++ packages/react-libraries/lib/buildJS.js
@@ -22,8 +22,17 @@ function logSuccess() {
   console.log(chalk.green('JavaScript build successful.'));
 }
 
+function formatError(error) {
+  return `\n${error.message}`;
+}
+
 function build() {
-  return Promise.all(config.map(buildConfig)).then(logSuccess);
+  return Promise.all(config.map(buildConfig))
+    .then(logSuccess)
+    .catch(e => {
+      console.error(formatError(e));
+      Promise.reject();
+    });
 }
 
 function start() {
@@ -37,7 +46,7 @@ function start() {
 
       case 'ERROR':
       case 'FATAL':
-        console.error('error:', event);
+        console.error(formatError(event.error));
         break;
 
       default: